+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Remove erroneous textures lump size limit
 - Add 'bsp entities' command
 - Allow map graph to read fro STDIN
+- Fix bsp entities displaying NUL byte
 
 # v1.6.1
 - Fix CI